#16 Home Page Styling

+++ src/app/settings/projects/projects.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, ChangeDetectionStrategy } from "@angular/core";
 import { ProjectsService } from "../../api/projects/projects.service";
+import { OrganizationsService } from "src/app/api/organizations/organizations.service";
 
 @Component({
   selector: "app-projects",
@@ -9,13 +10,21 @@ import { ProjectsService } from "../../api/projects/projects.service";
 })
 export class ProjectsComponent implements OnInit {
   projects$ = this.projectsService.getProjects;
-  constructor(private projectsService: ProjectsService) {}
+  activeOrganizationDetail$ = this.organizationsService
+    .activeOrganizationDetail$;
+
+  constructor(
+    private projectsService: ProjectsService,
+    private organizationsService: OrganizationsService
+  ) {}
 
   ngOnInit() {
     this.projectsService.retrieveProjects();
   }
 
   onDelete(projectId: number) {
-    this.projectsService.deleteProject(projectId).toPromise();
+    if (window.confirm("Are you sure you want to delete your project?")) {
+      this.projectsService.deleteProject(projectId).toPromise();
+    }
   }
 }
